feat: 项目出库-暂存

+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/entity/project/Project.java
@@ -234,7 +234,7 @@ public class Project extends BaseEntity{
      * 当前出库的日志id
      */
     @Column(name = "outbound_log_id", nullable = false )
-    private Integer outboundLogId;
+    private Long outboundLogId;
 
     @Transient
     private Integer completePercent;
