+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/entity/project/ProjectSimple.java
@@ -226,7 +226,7 @@ public class ProjectSimple extends BaseEntity {
      * 当前出库的日志id
      */
     @Column(name = "outbound_log_id", nullable = false )
-    private Integer outboundLogId;
+    private Long outboundLogId;
 
 /*    @OneToOne(fetch = FetchType.LAZY)
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
